refactor: change file.js naming
