+++ src/views/FileImport.vue
@@ -1,6 +1,6 @@
 <script setup>
 import GoBack from '@/components/go-back.vue'
-import { processCsv, processJson } from '@/utils/file.js'
+import { processCsv, processJson } from '@/utils/process-file-formats.js'
 import OptimizeRouteButton from '@/components/optimize-route-button.vue'
 
 const onChange = (e) => {
